Move KitCommand
